requirements.txt
# python 3.10.13
numpy
pandas
Pillow
scikit-learn
scipy
tqdm

torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0

